requirements.txt
# Core dependencies
selenium==4.17.2
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.0

# Database
sqlalchemy>=2.0.30

# Data processing
pandas==2.3.1
numpy==2.3.2

# Utilities
fake-useragent==1.4.0
undetected-chromedriver==3.5.4

# Logging
colorlog==6.8.0

# Monitoring and notifications
schedule==1.2.0
plyer==2.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Development
black==23.12.1
flake8==7.0.0